--- bounties_owed.csv
@@ -1,5 +1,0 @@
-PR,Developer,Feature,Bounty_Amount,Date_Merged,Status
-195,dhvll,Package Manager Wrapper,100,2025-11-18,PENDING
-198,aliraza556,Installation Rollback,150,2025-11-18,PENDING
-21,aliraza556,Config Templates,150,2025-11-18,PENDING
-17,chandrapratamar,Package Manager (original),100,2025-11-18,PENDING
