--- bounties_pending.json
@@ -1,1 +1,0 @@
-[]
