+++ cortex/installation_verifier.py
