+++ cortex/logging_system.py
