--- payments_history.json
@@ -1,1 +1,0 @@
-[]
