--- pr_status.json
@@ -1,1 +1,0 @@
-[{"author":{"id":"MDQ6VXNlcjg3MDY4MzM5","is_bot":false,"login":"aliraza556","name":"Ali Raza"},"isDraft":false,"mergeable":"MERGEABLE","number":198,"state":"OPEN","title":"Add installation history tracking and rollback support"},{"author":{"id":"MDQ6VXNlcjczMzc2NjM0","is_bot":false,"login":"mikejmorgan-ai","name":"Mike Morgan"},"isDraft":false,"mergeable":"MERGEABLE","number":197,"state":"OPEN","title":"Remove duplicate workflow"},{"author":{"id":"MDQ6VXNlcjU1MzE3NzY4","is_bot":false,"login":"dhvll","name":"Dhaval"},"isDraft":false,"mergeable":"MERGEABLE","number":195,"state":"OPEN","title":"feat: Introduce intelligent package manager wrapper"},{"author":{"id":"MDQ6VXNlcjY0MDI1NzYy","is_bot":false,"login":"AlexanderLuzDH","name":"Alexander Luz"},"isDraft":false,"mergeable":"CONFLICTING","number":38,"state":"OPEN","title":"feat: Add system requirements pre-flight checker (Issue #28)"},{"author":{"id":"MDQ6VXNlcjg3MDY4MzM5","is_bot":false,"login":"aliraza556","name":"Ali Raza"},"isDraft":false,"mergeable":"MERGEABLE","number":21,"state":"OPEN","title":"feat: Configuration File Template System - Generate nginx, PostgreSQL, Redis, Docker Compose & Apache configs"},{"author":{"id":"U_kgDOBw4eqA","is_bot":false,"login":"Sahilbhatane","name":"Sahil Bhatane"},"isDraft":true,"mergeable":"CONFLICTING","number":18,"state":"OPEN","title":"Add CLI interface for cortex command - Fixes #11"},{"author":{"id":"U_kgDOBwep0g","is_bot":false,"login":"chandrapratamar","name":"Chandra Pratama"},"isDraft":false,"mergeable":"CONFLICTING","number":17,"state":"OPEN","title":"Add Intelligent package manager wrapper"}]
