+++ scripts/audit_cortex_status.sh
