+++ scripts/cortex-master-automation.sh
