+++ scripts/cortex-master-quarterback.sh
