+++ scripts/cortex-master-update.sh
