+++ scripts/cortex-master.sh
