+++ scripts/setup-github-automation.sh
