+++ scripts/setup_and_upload.sh
