+++ test/test_context_memory.py
@@ -14,9 +14,9 @@ import sys
 import os
 
 # Add parent directory to path for imports
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
 
-from context_memory import (
+from cortex.context_memory import (
     ContextMemory,
     MemoryEntry,
     Pattern,
