+++ test/test_installation_verifier.py
@@ -4,7 +4,12 @@ Tests for Installation Verification System
 """
 
 import unittest
-from installation_verifier import (
+import sys
+import os
+
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
+
+from cortex.installation_verifier import (
     InstallationVerifier,
     VerificationStatus,
     VerificationTest
