+++ test/test_llm_router.py
@@ -13,9 +13,9 @@ import os
 import sys
 
 # Add parent directory to path
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
 
-from llm_router import (
+from cortex.llm_router import (
     LLMRouter,
     TaskType,
     LLMProvider,
@@ -244,7 +244,7 @@ class TestCostTracking(unittest.TestCase):
 class TestClaudeIntegration(unittest.TestCase):
     """Test Claude API integration."""
     
-    @patch('llm_router.Anthropic')
+    @patch('cortex.llm_router.Anthropic')
     def test_claude_completion(self, mock_anthropic):
         """Test Claude completion with mocked API."""
         # Mock response
@@ -276,7 +276,7 @@ class TestClaudeIntegration(unittest.TestCase):
         self.assertEqual(result.tokens_used, 150)
         self.assertGreater(result.cost_usd, 0)
     
-    @patch('llm_router.Anthropic')
+    @patch('cortex.llm_router.Anthropic')
     def test_claude_with_system_message(self, mock_anthropic):
         """Test Claude handles system messages correctly."""
         mock_content = Mock()
@@ -313,7 +313,7 @@ class TestClaudeIntegration(unittest.TestCase):
 class TestKimiIntegration(unittest.TestCase):
     """Test Kimi K2 API integration."""
     
-    @patch('llm_router.OpenAI')
+    @patch('cortex.llm_router.OpenAI')
     def test_kimi_completion(self, mock_openai):
         """Test Kimi K2 completion with mocked API."""
         # Mock response
@@ -348,7 +348,7 @@ class TestKimiIntegration(unittest.TestCase):
         self.assertEqual(result.tokens_used, 150)
         self.assertGreater(result.cost_usd, 0)
     
-    @patch('llm_router.OpenAI')
+    @patch('cortex.llm_router.OpenAI')
     def test_kimi_temperature_mapping(self, mock_openai):
         """Test Kimi K2 temperature is scaled by 0.6."""
         mock_message = Mock()
@@ -380,7 +380,7 @@ class TestKimiIntegration(unittest.TestCase):
         call_args = mock_client.chat.completions.create.call_args
         self.assertAlmostEqual(call_args.kwargs["temperature"], 0.6, places=2)
     
-    @patch('llm_router.OpenAI')
+    @patch('cortex.llm_router.OpenAI')
     def test_kimi_with_tools(self, mock_openai):
         """Test Kimi K2 handles tool calling."""
         mock_message = Mock()
@@ -422,8 +422,8 @@ class TestKimiIntegration(unittest.TestCase):
 class TestEndToEnd(unittest.TestCase):
     """End-to-end integration tests."""
     
-    @patch('llm_router.Anthropic')
-    @patch('llm_router.OpenAI')
+    @patch('cortex.llm_router.Anthropic')
+    @patch('cortex.llm_router.OpenAI')
     def test_complete_with_routing(self, mock_openai, mock_anthropic):
         """Test complete() method with full routing."""
         # Mock Kimi K2 (should be used for system operations)
@@ -457,8 +457,8 @@ class TestEndToEnd(unittest.TestCase):
         self.assertEqual(response.provider, LLMProvider.KIMI_K2)
         self.assertIn("Installing", response.content)
     
-    @patch('llm_router.Anthropic')
-    @patch('llm_router.OpenAI')
+    @patch('cortex.llm_router.Anthropic')
+    @patch('cortex.llm_router.OpenAI')
     def test_fallback_on_error(self, mock_openai, mock_anthropic):
         """Test fallback when primary provider fails."""
         # Mock Kimi K2 to fail
@@ -499,7 +499,7 @@ class TestEndToEnd(unittest.TestCase):
 class TestConvenienceFunction(unittest.TestCase):
     """Test the complete_task convenience function."""
     
-    @patch('llm_router.LLMRouter')
+    @patch('cortex.llm_router.LLMRouter')
     def test_complete_task_simple(self, mock_router_class):
         """Test simple completion with complete_task()."""
         # Mock router
@@ -519,7 +519,7 @@ class TestConvenienceFunction(unittest.TestCase):
         self.assertEqual(result, "Test response")
         mock_router.complete.assert_called_once()
     
-    @patch('llm_router.LLMRouter')
+    @patch('cortex.llm_router.LLMRouter')
     def test_complete_task_with_system_prompt(self, mock_router_class):
         """Test complete_task() includes system prompt."""
         mock_response = Mock()
